Not done yet, but here it is so far. -feature: Add default double click for hosts -feature: Add rows selector to code -feature: Clean up images -feature: Resolve some, not all image anomalies -feature: Provide a red tab for syslog -feature: Make menu's dynamic all the time -feature: Make Levels so that they are level++ by default
Todo's include:
-feature: Allow going to syslog from Graphs
-feature: Sort columns
-QA: various code junk
-QA: Cacti Group Copyright


git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti_plugins/syslog/trunk@184 67214365-cc29-0410-bfbe-843132f0f3d7

+++ syslog.sql
@@ -1,6 +1,6 @@
--- 
+--
 -- Table structure for table `syslog`
--- 
+--
 
 DROP TABLE IF EXISTS syslog;
 CREATE TABLE syslog (
@@ -10,7 +10,7 @@ CREATE TABLE syslog (
   `time` time default NULL,
   host varchar(128) default NULL,
   message text,
-  seq int(10) unsigned NOT NULL auto_increment,
+  seq bigint unsigned NOT NULL auto_increment,
   PRIMARY KEY  (seq),
   KEY `date` (`date`),
   KEY `time` (`time`),
@@ -22,9 +22,9 @@ CREATE TABLE syslog (
 
 -- --------------------------------------------------------
 
--- 
+--
 -- Table structure for table `syslog_alert`
--- 
+--
 DROP TABLE IF EXISTS syslog_alert;
 CREATE TABLE syslog_alert (
   id int(10) NOT NULL auto_increment,
@@ -40,9 +40,9 @@ CREATE TABLE syslog_alert (
 
 -- --------------------------------------------------------
 
--- 
+--
 -- Table structure for table `syslog_incoming`
--- 
+--
 
 DROP TABLE IF EXISTS syslog_incoming;
 CREATE TABLE syslog_incoming (
@@ -52,7 +52,7 @@ CREATE TABLE syslog_incoming (
   `time` time default NULL,
   host varchar(128) default NULL,
   message text,
-  seq int(10) unsigned NOT NULL auto_increment,
+  seq bigint unsigned NOT NULL auto_increment,
   `status` tinyint(4) NOT NULL default '0',
   PRIMARY KEY  (seq),
   KEY `status` (`status`)
@@ -60,9 +60,9 @@ CREATE TABLE syslog_incoming (
 
 -- --------------------------------------------------------
 
--- 
+--
 -- Table structure for table `syslog_remove`
--- 
+--
 
 DROP TABLE IF EXISTS syslog_remove;
 CREATE TABLE syslog_remove (
@@ -76,9 +76,9 @@ CREATE TABLE syslog_remove (
   PRIMARY KEY  (id)
 ) TYPE=MyISAM;
 
--- 
+--
 -- Table structure for table `syslog_reports`
--- 
+--
 
 DROP TABLE IF EXISTS syslog_reports;
 CREATE TABLE syslog_reports (
